Check deps and fix them in ExportCap function.
At this point, the library checks dependencies when adding. Later, when calling ExportCap, we corrected notes
to point relative to position within the .cap.

It might be better to change the JSON right away when parsing files the first time. But for now, it stays that way for testing.

I had problems with permissions when using `/tmp` on linux, for now the temporary directory is `.tmp/` in the same directory where the program runs.